add `editLink.component`
