+++ .changeset/pink-chicken-exist.md
@@ -1,0 +1,5 @@
+---
+'nextra-theme-docs': patch
+---
+
+add `editLink.component`
