+++ examples/swr-site/theme.config.tsx
@@ -103,9 +103,11 @@ const config: DocsThemeConfig = {
   darkMode: true,
   docsRepositoryBase:
     "https://github.com/shuding/nextra/blob/core/examples/swr-site",
-  editLinkText() {
-    const { locale } = useRouter();
-    return EDIT_TEXT[locale];
+  editLink: {
+    text() {
+      const { locale } = useRouter();
+      return EDIT_TEXT[locale];
+    },
   },
   feedback: {
     labels: "feedback",
