+++ packages/nextra-theme-docs/src/utils/index.ts
@@ -1,4 +1,5 @@
 export { getFSRoute } from './get-fs-route'
+export { getGitEditUrl } from './get-git-edit-url'
 export { getGitIssueUrl } from './get-git-issue-url'
 export { getHeadingText } from './get-heading-text'
 export { Item, PageItem, MenuItem, normalizePages } from './normalize-pages'
